second Step
